+++ src/project.h
@@ -93,6 +93,6 @@ const char *gf_blind_desc(int type);
 int gf_name_to_idx(const char *name);
 const char *gf_idx_to_name(int type);
 bool project(int who, int rad, int y, int x, int dam, int typ, int flg,
-			 int degrees_of_arc, byte diameter_of_source);
+			 int degrees_of_arc, int diameter_of_source);
 
 #endif /* !PROJECT_H */
